+++ Source/WebCore/SaferCPPExpectations/UncountedLocalVarsCheckerExpectations
@@ -116,6 +116,7 @@ style/RuleSetBuilder.cpp
 style/StyleBuilder.cpp
 style/StyleBuilderCustom.h
 style/StyleDifference.cpp
+style/StyleDocumentScope.cpp
 style/StyleExtractorCustom.h
 style/StyleInvalidationFunctions.h
 style/StyleRelations.cpp
