+++ Source/WebCore/dom/Element.cpp
@@ -150,13 +150,13 @@
 #include "ShadowRootInit.h"
 #include "SimulatedClick.h"
 #include "SlotAssignment.h"
+#include "StyleDocumentScope.h"
 #include "StyleableInlines.h"
 #include "StyleComputedStyle+SettersInlines.h"
 #include "StyleInvalidator.h"
 #include "StylePrimitiveNumericTypes+Evaluation.h"
 #include "StyleProperties.h"
 #include "StyleResolver.h"
-#include "StyleScope.h"
 #include "StyleTreeResolver.h"
 #include "StyleZoomPrimitivesInlines.h"
 #include "TextIterator.h"
