+++ Source/WebCore/dom/ExtensionStyleSheets.cpp
@@ -37,9 +37,9 @@
 #include "ProcessingInstruction.h"
 #include "SVGStyleElement.h"
 #include "Settings.h"
+#include "StyleDocumentScope.h"
 #include "StyleInvalidator.h"
 #include "StyleResolver.h"
-#include "StyleScope.h"
 #include "StyleSheetContents.h"
 #include "StyleSheetList.h"
 #include "UserContentController.h"
