+++ Source/WebCore/dom/ViewTransition.cpp
@@ -57,9 +57,9 @@
 #include "RenderView.h"
 #include "RenderViewTransitionCapture.h"
 #include "StyleComputedStyle+GettersInlines.h"
+#include "StyleDocumentScope.h"
 #include "StyleExtractor.h"
 #include "StyleResolver.h"
-#include "StyleScope.h"
 #include "StyleTransformFunction.h"
 #include "StyleZoomPrimitivesInlines.h"
 #include "Styleable.h"
