+++ Source/WebCore/page/LocalFrame.cpp
@@ -117,8 +117,8 @@
 #include "ServiceWorkerGlobalScope.h"
 #include "Settings.h"
 #include "StyleComputedStyle+GettersInlines.h"
+#include "StyleDocumentScope.h"
 #include "StyleProperties.h"
-#include "StyleScope.h"
 #include "TextNodeTraversal.h"
 #include "TextResourceDecoder.h"
 #include "UserContentController.h"
