+++ Source/WebCore/page/LocalFrameView.cpp
@@ -132,8 +132,8 @@
 #include "ShadowRoot.h"
 #include "SimpleRange.h"
 #include "StyleComputedStyle+SettersInlines.h"
+#include "StyleDocumentScope.h"
 #include "StyleResolver.h"
-#include "StyleScope.h"
 #include "TextIndicator.h"
 #include "TextIterator.h"
 #include "TextResourceDecoder.h"
