+++ Source/WebCore/style/StyleTreeResolver.cpp
@@ -65,11 +65,11 @@
 #include "StyleAdjuster.h"
 #include "StyleBuilder.h"
 #include "StyleComputedStyle+SettersInlines.h"
+#include "StyleDocumentScope.h"
 #include "StyleFontSizeFunctions.h"
 #include "StyleOriginatedTimelinesController.h"
 #include "StylePositionTryFallbackTactic.h"
 #include "StyleResolver.h"
-#include "StyleScope.h"
 #include "StyleTreeResolverInlines.h"
 #include "SVGElement.h"
 #include "Text.h"
