+++ Source/WebCore/xml/parser/XMLDocumentParser.cpp
@@ -51,7 +51,7 @@
 #include "SVGStyleElement.h"
 #include "ScriptElement.h"
 #include "ScriptSourceCode.h"
-#include "StyleScope.h"
+#include "StyleDocumentScope.h"
 #include "TextResourceDecoder.h"
 #include "XMLNSNames.h"
 #include <wtf/Ref.h>
